refactor: extract archive display component
